+++ HackerRank/c++/BasicDataTypes/BasicDataTypes.cpp
@@ -1,0 +1,15 @@
+#include <iostream>
+#include <cstdio>
+using namespace std;
+
+int main() {
+    // Complete the code.
+    char t;
+    int k;
+    long m;
+    double o;
+    float z;
+    scanf("%d %ld %c %f %lf",&k,&m,&t,&z,&o);
+    printf("%d\n%ld \n%c \n%f \n%lf",k,m,t,z,o);
+    return 0;
+}
