+++ HackerRank/c++/BoxIt/BoxIt.cpp
@@ -1,43 +1,12 @@
-// Design a class named Box whose dimensions are integers and private to the class. The dimensions are labelled: length , breadth , and height .
-
-// The default constructor of the class should initialize , , and to .
-
-// The parameterized constructor Box(int length, int breadth, int height) should initialize Box's and to length, breadth and height.
-
-// The copy constructor BoxBox ) should set and to 's and , respectively.
-
-// Apart from the above, the class should have functions:
-
-//     int getLength() - Return box's length
-//     int getBreadth() - Return box's breadth
-//     int getHeight() - Return box's height
-//     long long CalculateVolume() - Return the volume of the box
-
-// Overload the operator for the class Box. Box Box if:
-
-//     <
-//     < and ==
-//     < and == and ==
-
-// Overload operator for the class Box().
-// If is an object of class Box:
-
-// should print , and on a single line separated by spaces.
-
-// Constraints
-
-
-// Two boxes being compared using the operator will not have all three dimensions equal.
-
 #include<bits/stdc++.h>
 
 using namespace std;
 
 
-//Implement the class Box  
+//Implement the class Box
 //l,b,h are integers representing the dimensions of the box
 
-// The class should have the following functions : 
+// The class should have the following functions :
 class Box{
     private:
         int l,b,h;
@@ -60,7 +29,7 @@ class Box{
     int getHeight(){
             return l;
         }
-    
+
     Box(){
         l = 0;
         b = 0;
@@ -83,15 +52,15 @@ class Box{
     bool operator < (Box &B){
         if(this->l < B.l){ return true; }
     else if(this->l == B.l && this->b < B.b){ return true; }
-    else if(this->h < B.h && this->l == B.l && this->b == B.b){ return true; } 
+    else if(this->h < B.h && this->l == B.l && this->b == B.b){ return true; }
         return false; }
-    
+
     friend ostream& operator<< (ostream& output, const Box& B){
         output << B.l << " " << B.b << " " << B.h;
         return output;
     }
 };
-// Constructors: 
+// Constructors:
 // Box();
 // Box(int,int,int);
 // Box(Box);
