+++ HackerRank/c++/ClassAndObjects/ClassAndObjects.out
