+++ HackerRank/c++/ConditionalStatements/ConditionalStatements.cpp
@@ -1,0 +1,36 @@
+#include <bits/stdc++.h>
+
+using namespace std;
+
+
+
+int main()
+{
+    int n;
+    cin >> n;
+    cin.ignore(numeric_limits<streamsize>::max(), '\n');
+    if(n==1){
+        cout << "one";
+    }else if(n==2){
+        cout << "two";
+    }else if(n==3){
+        cout << "three";
+    }else if(n==4){
+        cout << "four";
+    }else if(n==5){
+        cout << "five";
+    }else if(n==6){
+        cout << "six";
+    }else if(n==7){
+        cout << "seven";
+    }else if(n==8){
+        cout << "eight";
+    }else if(n==9){
+        cout << "nine";
+    }else{
+        cout << "Greater than 9";
+    }
+    // Write Your Code Here
+
+    return 0;
+}
