+++ HackerRank/c++/InputandOutput/InputAndOutput.cpp
@@ -1,0 +1,16 @@
+#include <cmath>
+#include <cstdio>
+#include <vector>
+#include <iostream>
+#include <algorithm>
+using namespace std;
+
+
+int main() {
+    /* Enter your code here. Read input from STDIN. Print output to STDOUT */
+    int s,l,m,t;
+    cin >> s >> l >> m;
+    t=s+l+m;
+    cout << t;
+    return 0;
+}
