+++ HackerRank/c++/README.md
@@ -1,0 +1,3 @@
+### This codes working for hacker rank practice. Practice links:
+
+I will add link
