+++ HackerRank/c++/SayHello,World!WithC++/README.md
@@ -1,0 +1,19 @@
+[Say "Hello, World!" With C++](https://www.hackerrank.com/challenges/cpp-hello-world/problem)
+
+
+Objective
+This is a simple challenge to help you practice printing to stdout. You may also want to complete Solve Me First in C++ before attempting this challenge.
+
+We're starting out by printing the most famous computing phrase of all time! In the editor below, use either printf or cout to print the string to stdout.
+
+Input Format
+
+You do not need to read any input in this challenge.
+
+Output Format
+
+Print to stdout.
+
+Sample Output
+
+Hello, World!
