+++ HackerRank/c++/SayHello,World!WithC++/SayHello,World!WithC++.cpp
@@ -1,0 +1,8 @@
+#include <iostream>
+#include <cstdio>
+using namespace std;
+
+int main() {
+	printf("Hello, World!");
+	return 0;
+}
