+++ HackerRank/c++/StringStream/StringStream.out
