+++ HackerRank/c++/Strings/Strings.out
