--- HackerRank/c++/c1.cpp
@@ -1,20 +1,0 @@
-#include <cmath>
-#include <cstdio>
-#include <vector>
-#include <iostream>
-#include <algorithm>
-using namespace std;
-
-vector<int>a;
-vector<int>b;
-int main() {
-	int l,m,input;
-	/* Enter your code here. Read input from STDIN. Print output to STDOUT */
-	//cin >> l >> m;
-	while (cin >> input || a[0] != a.size()){
-		a.push_back(input);
-	}
-	cout << a[0];
-			        
-	return 1;
-}
