--- HackerRank/c++/c1.cpp~
@@ -1,18 +1,0 @@
-include <iostream>
-#include <vector>
-using namespace std;
-vector<int>dizi;                                     // VEKTOR TANIMLANDI
-int main()
-{
-    for(int i=0;i<10;i++){                           // VEKTORUN İÇİNE 0-10 A KADAR DEĞERLER ALDI
-        dizi.push_back(i);
-}
-    for(int i=0;i<dizi.size();i++){                  // VEKTÜRUN ELEMANLARI YAZILDI
-        cout << dizi.at(i) << endl;
-}
-        dizi.pop_back();                             // VEKTORUN SON ELEMANI SİLİNDİ
-        cout<<"----------------------"<<endl;
-        cout << dizi.back() << endl;                 // VEKTORÜN SON ELEMANI YAZILDI
-        cout<<"----------------------"<<endl;
-    return 0;
-    }
